+++ logfire/_internal/logs.py
@@ -17,9 +17,9 @@ class ProxyLoggerProvider(LoggerProvider):
 
     provider: LoggerProvider
 
-    loggers: WeakSet[ProxyLogger] = dataclasses.field(default_factory=WeakSet)
+    loggers: WeakSet[ProxyLogger] = dataclasses.field(default_factory=WeakSet)  # type: ignore[reportUnknownVariableType]
     lock: Lock = dataclasses.field(default_factory=Lock)
-    suppressed_scopes: set[str] = dataclasses.field(default_factory=set)
+    suppressed_scopes: set[str] = dataclasses.field(default_factory=set)  # type: ignore[reportUnknownVariableType]
 
     def get_logger(
         self, name: str, version: str | None = None, schema_url: str | None = None, attributes: Attributes | None = None