+++ logfire/_internal/tracer.py
@@ -53,9 +53,9 @@ class ProxyTracerProvider(TracerProvider):
 
     provider: TracerProvider
     config: LogfireConfig
-    tracers: WeakKeyDictionary[_ProxyTracer, Callable[[], Tracer]] = field(default_factory=WeakKeyDictionary)
+    tracers: WeakKeyDictionary[_ProxyTracer, Callable[[], Tracer]] = field(default_factory=WeakKeyDictionary)  # type: ignore[reportUnknownVariableType]
     lock: Lock = field(default_factory=Lock)
-    suppressed_scopes: set[str] = field(default_factory=set)
+    suppressed_scopes: set[str] = field(default_factory=set)  # type: ignore[reportUnknownVariableType]
 
     def set_provider(self, provider: SDKTracerProvider) -> None:
         with self.lock:
@@ -286,7 +286,7 @@ class PendingSpanProcessor(SpanProcessor):
             return
 
         attributes = span.attributes
-        if not attributes or attributes.get(ATTRIBUTES_SPAN_TYPE_KEY) == 'log':
+        if not attributes or attributes.get(ATTRIBUTES_SPAN_TYPE_KEY) not in (None, 'span'):
             return
 
         real_span_context = span.get_span_context()